requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
yfinance
boto3

apache-airflow-providers-amazon
apache-airflow-providers-common-io